tests: add tests for general resources
