+++ src/rds-monitoring-mcp-server/awslabs/rds_monitoring_mcp_server/common/utils.py
@@ -16,7 +16,6 @@
 
 from datetime import datetime
 from loguru import logger
-from pathlib import Path
 from typing import Any, Optional
 
 
@@ -95,25 +94,3 @@ def convert_string_to_datetime(default: datetime, date_string: Optional[str] = N
     except ValueError as e:
         logger.warning(f"Error parsing end_date '{date_string}': {str(e)}. Using default value.")
         return default
-
-
-def load_markdown_file(filename: str) -> str:
-    """Load a markdown file from the static/react directory.
-
-    Args:
-        filename (str): The name of the markdown file to load (e.g. 'basic-ui-setup.md')
-
-    Returns:
-        str: The content of the markdown file, or empty string if file not found
-    """
-    base_dir = Path(__file__).parent.parent
-    static_dir = base_dir / 'static'
-    file_path = static_dir / filename
-
-    if file_path.exists():
-        with open(file_path, 'r', encoding='utf-8') as f:
-            logger.info(f'Loading markdown file: {file_path}')
-            return f.read()
-    else:
-        logger.error('File not found: {file_path}')
-        return f'Warning: File not found: {file_path}'
